# api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Task Manager API
  version: 1.0.0
  description: Full-stack Task Management Application API

servers:
  - url: http://localhost:8080
    description: Local development

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [username, email, password, fullName]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        fullName:
          type: string

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        userId:
          type: string
        username:
          type: string
        email:
          type: string
        fullName:
          type: string

    TaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [TODO, IN_PROGRESS, IN_REVIEW, DONE, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        dueDate:
          type: string
          format: date
        assigneeId:
          type: string

    TaskResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: string
        assigneeId:
          type: string
        creatorId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    DashboardStats:
      type: object
      properties:
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        pendingTasks:
          type: integer
        overdueTasks:
          type: integer
        tasksByStatus:
          type: object
          additionalProperties:
            type: integer
        tasksByPriority:
          type: object
          additionalProperties:
            type: integer

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        message:
          type: string

paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error or duplicate user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login with credentials
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/tasks:
    get:
      tags: [Tasks]
      summary: Get all tasks for the current user
      responses:
        '200':
          description: List of tasks sorted by due date and priority
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized

    post:
      tags: [Tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
        '401':
          description: Unauthorized

  /api/tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [Tasks]
      summary: Get a task by ID
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Task not found
        '401':
          description: Unauthorized

    put:
      tags: [Tasks]
      summary: Update a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Task not found or validation error
        '401':
          description: Unauthorized

    delete:
      tags: [Tasks]
      summary: Delete a task
      responses:
        '204':
          description: Task deleted
        '400':
          description: Task not found
        '401':
          description: Unauthorized
        '403':
          description: Not authorized to delete this task

  /api/tasks/dashboard/stats:
    get:
      tags: [Tasks]
      summary: Get dashboard statistics
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          description: Unauthorized